backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
colorama==0.4.6
cryptography==45.0.4
distro==1.9.0
Django==5.2.3
django-appconf==1.1.0
django-cors-headers==4.7.0
django-cryptography-django5==2.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.90.0
pdfkit==1.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
python-decouple==3.8
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
